pyproject.toml
[project]
name = "mcp-server-access-mdb"
version = "0.1.0"
description = "MCP server for Microsoft Access databases (.mdb)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chardet>=5.2.0",
    "fastmcp>=2.8.1",
    "pandas>=2.3.0",
    "sqlalchemy-access>=1.1.3",
]
